// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const TourType = {
  guided: 'guided',
  self_guided: 'self_guided',
  rental: 'rental'
} as const

export type TourType = (typeof TourType)[keyof typeof TourType]


export const Difficulty = {
  easy: 'easy',
  moderate: 'moderate',
  challenging: 'challenging'
} as const

export type Difficulty = (typeof Difficulty)[keyof typeof Difficulty]


export const BookingStatus = {
  pending: 'pending',
  rejected: 'rejected',
  paid: 'paid'
} as const

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus]


export const RidingExperience = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  experienced: 'experienced'
} as const

export type RidingExperience = (typeof RidingExperience)[keyof typeof RidingExperience]


export const AdminRole = {
  super_admin: 'super_admin',
  staff: 'staff'
} as const

export type AdminRole = (typeof AdminRole)[keyof typeof AdminRole]


export const VehicleType = {
  ural_sidecar: 'ural_sidecar',
  dnepr_sidecar: 'dnepr_sidecar',
  solo_bike: 'solo_bike'
} as const

export type VehicleType = (typeof VehicleType)[keyof typeof VehicleType]